Feature/admin

+++ src/services/teams.service.js
@@ -2,7 +2,7 @@ export class TeamsService {
   constructor(teamsRepository) {
     this.teamsRepository = teamsRepository;
   }
-  // role 체크 메서드
+
   checkAdminRole = async (userId) => {
     const isAdmin = await this.teamsRepository.getUserByUserId(userId);
     if (!isAdmin) {
@@ -15,7 +15,11 @@ export class TeamsService {
   };
 
   getAllTeams = async (orderKey, orderValue, projectId) => {
-    const teams = await this.teamsRepository.getAllTeams(orderKey, orderValue, projectId);
+    const teams = await this.teamsRepository.getAllTeams(
+      orderKey,
+      orderValue,
+      projectId,
+    );
     return teams;
   };
 
@@ -26,33 +30,31 @@ export class TeamsService {
   };
 
   createTeam = async (userId, projectId, name, memberList) => {
-    //Parameter - user.role이 admin인지 검증하기
     await this.checkAdminRole(userId);
 
     const project = await this.teamsRepository.getProjectByProjectId(projectId);
     if (!project) throw new Error("존재하지 않는 프로젝트 입니다.");
 
-    //레파지토리 계층에 클래스 생성 요청
+    //팀 생성 요청
     const createdTeam = await this.teamsRepository.createTeam(
       name,
       projectId,
       memberList,
     );
-    //Return
     return createdTeam;
   };
 
   updateTeam = async (userId, teamId, projectId, name, memberList) => {
-    //Parameter - user.role이 admin인지 검증하기
     await this.checkAdminRole(userId);
 
     const project = await this.teamsRepository.getProjectByProjectId(projectId);
     if (!project) throw new Error("존재하지 않는 프로젝트 입니다.");
 
-    // 클래스 존재 확인
+    //팀 존재 확인
     await this.getTeamByTeamId(teamId);
+    if (!teamId) throw new Error("존재하지 않는 팀입니다.");
 
-    //레파지토리 계층에 클래스 생성 요청
+    //팀 수정 요청
     const updatedTeam = await this.teamsRepository.updateTeam(
       teamId,
       projectId,
@@ -60,18 +62,17 @@ export class TeamsService {
       memberList,
     );
 
-    //Return
     return updatedTeam;
   };
 
   deleteTeam = async (userId, teamId) => {
-    //Parameter - user.role이 admin인지 검증하기
     await this.checkAdminRole(userId);
 
-    // 클래스 존재 확인
+    // 팀 존재 확인
     await this.getTeamByTeamId(teamId);
+    if (!teamId) throw new Error("존재하지 않는 팀입니다.");
 
-    //레파지토리 계층에 클래스 생성 요청
+    // 팀 삭제 요청
     return await this.teamsRepository.deleteTeam(teamId);
   };
 }
